Refactor image comparison utils
