+++ projects/UnitTest/UnitTest.vcxproj.filters
@@ -86,12 +86,18 @@
     <ClCompile Include="src\render\TestImageComparison.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="src\render\test_render_utils.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\rhi\test_rhi_utils.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\render\test_render_utils.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
